+++ met4j-graph/src/main/java/fr/inrae/toulouse/metexplore/met4j_graph/computation/connect/SteinerTreeApprox.java
@@ -42,6 +42,7 @@ import java.util.List;
 import java.util.Set;
 
 import fr.inrae.toulouse.metexplore.met4j_graph.core.BioGraph;
+import fr.inrae.toulouse.metexplore.met4j_graph.core.GraphFactory;
 import org.jgrapht.alg.spanning.KruskalMinimumSpanningTree;
 import org.jgrapht.graph.DirectedWeightedMultigraph;
 
@@ -57,13 +58,14 @@ import fr.inrae.toulouse.metexplore.met4j_core.biodata.BioEntity;
  * @version $Id: $Id
  */
 public class SteinerTreeApprox<V extends BioEntity, E extends Edge<V>, G extends BioGraph<V ,E>>{
-	
+
 	/** The graph. */
 	public final G g;
-	
+
 	/** if the graph should be treated as weighted or not **/
 	public boolean weighted = true;
 	public boolean undirected = false;
+	public boolean pruning = true;
 
 	/**
 	 * Instantiates a new steiner tree computor.
@@ -73,32 +75,28 @@ public class SteinerTreeApprox<V extends BioEntity, E extends Edge<V>, G extends
 	public SteinerTreeApprox(G g) {
 		this.g=g;
 	}
-	
+
 	/**
 	 * Instantiates a new steiner tree computor.
-	 *
 	 * @param g the graph
-	 * @param weighted if the graph should be treated as weighted
+	 * @param useWeights consider edge weights sum rather than path length (default = true)
+	 * @param useEdgeDirection consider edge direction in path search (default = true)
+	 * @param prune prune union of shortest paths to remove cycle (default = true)
 	 */
-	public SteinerTreeApprox(G g, boolean weighted) {
+	public SteinerTreeApprox(G g, boolean useWeights, boolean useEdgeDirection, boolean prune) {
 		this.g=g;
-		this.weighted=weighted;
+		this.weighted = useWeights;
+		this.undirected = !useEdgeDirection;
+		this.pruning = prune;
 	}
 
-	public SteinerTreeApprox(G g, boolean weighted, boolean directed) {
-		this.g=g;
-		this.weighted=weighted;
-		this.undirected=!directed;
-	}
-	
 	/**
-	 * Gets the steiner tree list.
-	 *
+	 * Gets the lightest union of shortest paths connecting all nodes in set
+	 * Which correspond to the minimum spanning tree of the metric closure graph
 	 * @param terminal the targets list
-	 * @param weighted if the graph is weighted
 	 * @return the steiner tree list
 	 */
-	public List<E> getSteinerTreeList(Set<V> terminal, boolean weighted){
+	public List<E> getLightestUnionOfShortestPaths(Set<V> terminal){
 		Collection<V> unfound = new HashSet<>();
 		for(V v:terminal){
 			if(!g.containsVertex(v)) {
@@ -107,7 +105,7 @@ public class SteinerTreeApprox<V extends BioEntity, E extends Edge<V>, G extends
 			}
 		}
 		terminal.removeAll(unfound);
-		
+
 		ArrayList<E> list = new ArrayList<>();
 		CompressedGraph<V, E, G> cg = (new ShortestPath<>(g,!undirected)).getMetricClosureGraph(terminal, terminal, weighted);
 		KruskalMinimumSpanningTree<V, PathEdge<V,E>> kruskal = new KruskalMinimumSpanningTree<>(cg);
@@ -117,16 +115,16 @@ public class SteinerTreeApprox<V extends BioEntity, E extends Edge<V>, G extends
 		}
 		return list;
 	}
-	
+
 	/**
-	 * Gets the steiner tree list.
+	 * Gets the lightest union of shortest paths connecting all nodes in set
+	 * Which correspond to the minimum spanning tree of the metric closure graph
 	 *
 	 * @param startNodes the targets list
 	 * @param endNodes the targets list
-	 * @param weighted if the graph is weighted
 	 * @return the steiner tree list
 	 */
-	public List<E> getSteinerTreeList(Set<V> startNodes, Set<V> endNodes, boolean weighted){
+	public List<E> getLightestUnionOfShortestPaths(Set<V> startNodes, Set<V> endNodes){
 		Collection<V> unfound = new HashSet<>();
 		for(V v:startNodes){
 			if(!g.containsVertex(v)) {
@@ -142,7 +140,7 @@ public class SteinerTreeApprox<V extends BioEntity, E extends Edge<V>, G extends
 			}
 		}
 		endNodes.removeAll(unfound);
-		
+
 		ArrayList<E> list = new ArrayList<>();
 		DirectedWeightedMultigraph<V, PathEdge<V,E>> cg = (new ShortestPath<>(g,!undirected)).getMetricClosureGraph(startNodes, endNodes, weighted);
 		KruskalMinimumSpanningTree<V, PathEdge<V,E>> kruskal = new KruskalMinimumSpanningTree<>(cg);
@@ -151,6 +149,38 @@ public class SteinerTreeApprox<V extends BioEntity, E extends Edge<V>, G extends
 			list.addAll(edge.getPath().getEdgeList());
 		}
 		return list;
+	}
+
+	/**
+	 * Gets the steiner tree.
+	 *
+	 * @param startNodes the targets list
+	 * @param endNodes the targets list
+	 * @return the steiner tree
+	 */
+	public G getSteinerTree(Set<V> startNodes, Set<V> endNodes, GraphFactory<V,E,G> graphFactory){
+		List<E> edgeList = getLightestUnionOfShortestPaths(startNodes, endNodes);
+		G g2 = graphFactory.createGraphFromEdgeList(edgeList);
+		if(pruning) pruning(g2);
+		return g2;
+	}
+
+	/**
+	 * Gets the steiner tree.
+	 *
+	 * @param targetNodes the targets list
+	 * @return the steiner tree
+	 */
+	public G getSteinerTree(Set<V> targetNodes, GraphFactory<V,E,G> graphFactory){
+		return getSteinerTree(targetNodes,targetNodes,graphFactory);
+	}
+
+	private void pruning(G g2){
+		KruskalMinimumSpanningTree<V, E> kruskal = new KruskalMinimumSpanningTree<>(g2);
+		Set<E> mst = kruskal.getSpanningTree().getEdges();
+		Set<E> edges = new HashSet<>(g2.edgeSet());
+		edges.removeAll(mst);
+		g2.removeAllEdges(edges);
 	}
 
 }