+++ met4j-toolbox/src/main/java/fr/inrae/toulouse/metexplore/met4j_toolbox/generic/EdgeWeighting.java
@@ -39,10 +39,12 @@ public abstract class EdgeWeighting extends AbstractMet4jApplication {
     public String inputPath = null;
 
     @ParameterType(name= EnumParameterTypes.InputFile)
+    @Format(name= EnumFormats.Txt)
     @Option(name = "-sc", usage = "input Side compound file", required = false)
     public String inputSide = null;
 
     @ParameterType(name= EnumParameterTypes.OutputFile)
+    @Format(name= EnumFormats.Tsv)
     @Option(name = "-o", usage = "output edge weight file", required = true)
     public String outputPath = null;